requirements.txt
fastapi
uvicorn
pydantic
databases[postgresql]
python-dotenv
